+++ modules/dmrpp_module/DmrppCommon.h
@@ -76,7 +76,7 @@ private:
 	bool d_shuffle;
 	bool d_compact;
 	std::string d_byte_order;
-	std::vector<unsigned int> d_chunk_dimension_sizes;
+	std::vector<unsigned long long> d_chunk_dimension_sizes;
 	std::vector<std::shared_ptr<Chunk>> d_chunks;
 	bool d_twiddle_bytes;
 
@@ -154,7 +154,7 @@ public:
         return d_chunks;
     }
 
-    virtual const std::vector<unsigned int> &get_chunk_dimension_sizes() const {
+    virtual const std::vector<unsigned long long> &get_chunk_dimension_sizes() const {
     	return d_chunk_dimension_sizes;
     }
 
@@ -200,7 +200,7 @@ public:
             unsigned long long size,  unsigned long long offset, const std::string &position_in_array = "");
 
     virtual unsigned long add_chunk(const std::string &data_url, const std::string &byte_order,
-            unsigned long long size, const unsigned long long offset, const std::vector<unsigned int> &position_in_array);
+            unsigned long long size, const unsigned long long offset, const std::vector<unsigned long long> &position_in_array);
 
     virtual void dump(std::ostream & strm) const;
 };
